Corrected I2C datatypes

+++ src/Zanshin_BME680.cpp
@@ -22,7 +22,7 @@ BME680_Class::~BME680_Class() {}                                              //
 bool BME680_Class::begin() {                                                  // Find I2C device                  //
   begin(I2C_STANDARD_MODE);                                                   // Initialize I2c with slow speed   //
 } // of method begin()                                                        //                                  //
-bool BME680_Class::begin(const uint16_t i2cSpeed) {                           // Find I2C device                  //
+bool BME680_Class::begin(const uint32_t i2cSpeed) {                           // Find I2C device                  //
   Wire.begin();                                                               // Start I2C as master device       //
   Wire.setClock(i2cSpeed);                                                    // Set I2C bus speed                //
   for(_I2CAddress=0x76;_I2CAddress<=0x77;_I2CAddress++) {                     // loop all possible addresses      //
